+++ test/unit/test-helpers.js
@@ -125,12 +125,11 @@ const TestHelpers = {
   },
 
   /**
-   * Triggers an event on a DOM node natively.
+   * Creates an event.
    *
-   * @param  {Element} element
-   * @param  {string} eventType
+   * @param {string} eventType
    */
-  triggerDomEvent(element, eventType) {
+  createEvent(eventType) {
     let event;
 
     if (document.createEvent) {
@@ -142,6 +141,18 @@ const TestHelpers = {
     }
 
     event.eventName = eventType;
+
+    return event;
+  },
+
+  /**
+   * Triggers an event on a DOM node natively.
+   *
+   * @param  {Element} element
+   * @param  {string} eventType
+   */
+  triggerDomEvent(element, eventType) {
+    const event = TestHelpers.createEvent(eventType);
 
     if (document.createEvent) {
       element.dispatchEvent(event);