+++ test/unit/tracks/text-track.test.js
@@ -503,7 +503,6 @@ QUnit.test('stops processing if vttjs loading errored out', function(assert) {
   const errorSpy = sinon.spy();
   const oldVTT = window.WebVTT;
   const oldLogError = log.error;
-  const parserCreated = false;
   const reqs = [];
 
   window.xhr.onCreate = function(req) {
@@ -529,20 +528,14 @@ QUnit.test('stops processing if vttjs loading errored out', function(assert) {
 
   reqs.pop().respond(200, null, 'WEBVTT\n');
 
-  assert.ok(!parserCreated, 'WebVTT is not loaded, do not try to parse yet');
+  testTech.trigger('vttjserror');
+
+  assert.equal(errorSpy.callCount, 1, 'vttjs failed to load, so log.error was called');
 
   testTech.trigger('vttjserror');
-  const offSpyCall = testTech.off.getCall(0);
 
-  assert.ok(errorSpy.called, 'vttjs failed to load, so log.error was called');
-  if (errorSpy.called) {
-    assert.ok(
-      /^vttjs failed to load, stopping trying to process/.test(errorSpy.getCall(0).args[0]),
-      'log.error was called with the expected message'
-    );
-  }
-  assert.ok(!parserCreated, 'WebVTT is not loaded, do not try to parse yet');
-  assert.ok(offSpyCall, 'tech.off was called');
+  // vttjserror not called again
+  assert.equal(errorSpy.callCount, 1, 'vttjserror handler not called again');
 
   clock.restore();
   window.WebVTT = oldVTT;